+++ lib/screens/confirm_booking.dart
@@ -36,7 +36,7 @@ class _ConfirmBookingState extends State<ConfirmBooking> {
   String shiftBitstr1="",shiftBitstr2="",shiftBitstr3="";
   String shiftTextstr1,shiftTextstr2,shiftTextstr3;
   String takatext1,takatext2,takatext3;
-  int total_cost= 0;
+  double total_cost= 0;
   _ConfirmBookingState(this.convention,this.searchstring,this.dayString,this.selectedDate,this.secDate,this.thDate,this.dayOneShift, this.dayTwoShift, this.dayThreeShift);
   @override
   void dispose() {
@@ -388,7 +388,7 @@ class _ConfirmBookingState extends State<ConfirmBooking> {
             takatext1+=convention.mxprice.toString();
             takatext1 += "\u09F3";
             shiftTextstr1 += shft[i];
-            total_cost += int.parse(convention.mxprice);
+            total_cost += double.parse(convention.mxprice);
           }
         else shiftBitstr1 +=  "0";
         if(dayTwoShift[i]==true)
@@ -399,7 +399,7 @@ class _ConfirmBookingState extends State<ConfirmBooking> {
           takatext2+=convention.mxprice.toString();
           takatext2 += "\u09F3";
           shiftTextstr2 += shft[i];
-          total_cost += int.parse(convention.mxprice);
+          total_cost += double.parse(convention.mxprice);
         }
         else shiftBitstr2 += "0'";
         if(dayThreeShift[i]==true)
@@ -410,7 +410,7 @@ class _ConfirmBookingState extends State<ConfirmBooking> {
           takatext3+=convention.mxprice.toString();
           takatext3 += "\u09F3";
           shiftTextstr3 += shft[i];
-          total_cost += int.parse(convention.mxprice);
+          total_cost += double.parse(convention.mxprice);
         }
         else shiftBitstr3 += "0";
         print(shiftBitstr1);
